修改 lift 操作体系
